Move tables into a folder
