+++ static/app/views/performance/http/httpDomainSummaryPage.tsx
@@ -27,16 +27,16 @@ import {DurationChart} from 'sentry/views/performance/http/charts/durationChart'
 import {ResponseRateChart} from 'sentry/views/performance/http/charts/responseRateChart';
 import {ThroughputChart} from 'sentry/views/performance/http/charts/throughputChart';
 import {DomainStatusLink} from 'sentry/views/performance/http/domainStatusLink';
-import {
-  DomainTransactionsTable,
-  isAValidSort,
-} from 'sentry/views/performance/http/domainTransactionsTable';
 import {HTTPSamplesPanel} from 'sentry/views/performance/http/httpSamplesPanel';
 import {
   MODULE_TITLE,
   NULL_DOMAIN_DESCRIPTION,
   RELEASE_LEVEL,
 } from 'sentry/views/performance/http/settings';
+import {
+  DomainTransactionsTable,
+  isAValidSort,
+} from 'sentry/views/performance/http/tables/domainTransactionsTable';
 import {MetricReadout} from 'sentry/views/performance/metricReadout';
 import * as ModuleLayout from 'sentry/views/performance/moduleLayout';
 import {ModulePageProviders} from 'sentry/views/performance/modulePageProviders';
