+++ static/app/views/performance/http/tables/domainsTable.tsx
@@ -14,8 +14,8 @@ import {RATE_UNIT_TITLE, RateUnit} from 'sentry/utils/discover/fields';
 import {VisuallyCompleteWithData} from 'sentry/utils/performanceForSentry';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
-import {DomainCell} from 'sentry/views/performance/http/domainCell';
-import {ProjectIdCell} from 'sentry/views/performance/http/projectIdCell';
+import {DomainCell} from 'sentry/views/performance/http/tables/domainCell';
+import {ProjectIdCell} from 'sentry/views/performance/http/tables/projectIdCell';
 import {renderHeadCell} from 'sentry/views/starfish/components/tableCells/renderHeadCell';
 import type {MetricsResponse} from 'sentry/views/starfish/types';
 import {QueryParameterNames} from 'sentry/views/starfish/views/queryParameters';
